refactor(amf-deserializer): moved source files to src directory
