+++ packages/io/amf-deserializer/package.json
@@ -3,7 +3,7 @@
   "version": "2.1.3",
   "description": "AMF Deserializer for JSCAD",
   "repository": "https://github.com/jscad/OpenJSCAD.org/",
-  "main": "index.js",
+  "main": "src/index.js",
   "scripts": {
     "coverage": "nyc --all --reporter=html --reporter=text npm test",
     "test": "ava --verbose --timeout 2m 'tests/**/*.test.js'"
