+++ packages/io/amf-deserializer/src/constants.js
