+++ packages/io/amf-deserializer/src/helpers.js
